Support for triggering a full view update through NSFRCUpdateHandler.
This functionality is needed to make it possible to replace the predicate of the NSFetchRequest driving the NSFetchedResultsController, for instance when implementing a search feature. Changes to the fetch request demands that NSFRC's performFetch() is called anew, which in turn must be followed by a full reloadData() of the UITable- or CollectionView it drives.